+++ main.py
@@ -1,4 +1,5 @@
-from libs.helpers import define_args, dates_diff, format_number, format_bool
+from libs.helpers import define_args, dates_diff, format_number, get_test_stocks
+from libs.criteria import met_conditions_bullish
 from libs.stocktools import (
     get_asx_symbols,
     get_stock_data,
@@ -7,10 +8,9 @@ from libs.stocktools import (
     get_industry_from_web_batch
 )
 from libs.db import bulk_add_stocks, create_stock_table, delete_all_stocks, get_stocks, get_update_date
-from libs.settings import price_min, price_max, overextended_threshold_percent, minimum_volume_level
+from libs.settings import price_min, price_max, minimum_volume_level
 from libs.techanalysis import td_indicators, MA
 import pandas as pd
-import numpy as np
 
 
 def update_stocks():
@@ -27,119 +27,12 @@ def check_update_date():
     diff = dates_diff(last_update_date)
     if diff > 5:
         print("Stocks list updated more than 5 days ago, please run the --update first")
+        exit(0)
 
 
 def last_volume_5D_MA(volume_daily):
     volume_ma_20 = MA(volume_daily, 20, colname="volume")
     return volume_ma_20["ma20"].iloc[-1]
-
-
-def met_conditions_bullish(ohlc_with_indicators_daily,
-                           volume_daily,
-                           ohlc_with_indicators_weekly,
-                           consider_volume_spike=True,
-                           output=True):
-    # Checks if price action meets conditions
-    # Rules: MAs trending up, fast above slow, bullish TD count, volume spike
-    daily_condition_close_higher = (  # closes higher
-            ohlc_with_indicators_daily["close"].iloc[-1]
-            > ohlc_with_indicators_daily["close"].iloc[-2]
-    )
-    daily_condition_td = (  # bullish TD count
-            ohlc_with_indicators_daily["td_direction"].iloc[-1] == "green"
-    )
-    weekly_condition_td = (  # bullish TD count
-            ohlc_with_indicators_weekly["td_direction"].iloc[-1] == "green"
-    )
-
-    # MA check
-    # Used to have MA30, but it is not super helpful
-    ma_10 = MA(ohlc_with_indicators_daily, 10)
-    ma_50 = MA(ohlc_with_indicators_daily, 50)
-    ma_200 = MA(ohlc_with_indicators_daily, 200)
-
-    # MA 200 or even 50 may be None for too new stocks
-    ma_200_nan = np.isnan(ma_200["ma200"].iloc[-1])
-    ma_50_nan = np.isnan(ma_50["ma50"].iloc[-1])
-
-    if not ma_200_nan:
-        ma_consensio = (
-                (ma_50["ma50"].iloc[-1] > ma_200["ma200"].iloc[-1])
-                and (ma_10["ma10"].iloc[-1] > ma_50["ma50"].iloc[-1])
-        )
-    else:
-        ma_consensio = True
-        print('-- note: MA200 is NaN, the stock is too new')
-
-    # However, skip if there is no MA50 available to check
-    if ma_50_nan:
-        ma_consensio = True
-        print('-- excluding as MA50 is NaN')
-
-    # Volume MA and volume spike over the considered day
-    if consider_volume_spike:
-        volume_ma_20 = MA(volume_daily, 20, colname="volume")
-        mergedDf = volume_daily.merge(volume_ma_20, left_index=True, right_index=True)
-        mergedDf.dropna(inplace=True, how='any')
-        mergedDf["volume_above_average"] = mergedDf['volume'].ge(mergedDf['ma20'])  # GE is greater or equal
-        volume_condition = bool(mergedDf["volume_above_average"].iloc[-1])
-    else:
-        volume_condition = True
-
-    # All MA except for MA10 are rising
-    if not ma_200_nan:
-        ma_rising = (
-                (ma_50["ma50"].iloc[-1] >= ma_50["ma50"].iloc[-5])
-                and (ma_200["ma200"].iloc[-1] >= ma_200["ma200"].iloc[-5])
-        )
-    else:
-        ma_rising = (
-            (ma_50["ma50"].iloc[-1] >= ma_50["ma50"].iloc[-5])
-        )
-
-    # Close for the last week is not more than X% from the 4 weeks ago
-    not_overextended = (
-            ohlc_with_indicators_weekly["close"].iloc[-1]
-            < (1 + overextended_threshold_percent / 100) * ohlc_with_indicators_weekly["close"].iloc[-4]
-    )
-
-    # Last candle should actually be green (close above open)
-    last_candle_is_green = ohlc_with_indicators_daily['close'].iloc[-1] > ohlc_with_indicators_daily['open'].iloc[-1]
-
-    # Most recent close should be above the bodies of 10 candles prior
-    ohlc_with_indicators_daily['candle_body_upper'] = ohlc_with_indicators_daily[['open', 'close']].max(axis=1)
-    close_most_recent = float(ohlc_with_indicators_daily['close'].iloc[-1])
-    ohlc_with_indicators_daily["lower_than_recent"] = ohlc_with_indicators_daily['candle_body_upper'].lt(
-        close_most_recent)  # LT is lower than
-    # Do not include the most recent itself in the calculation. Take 10 previous before that.
-    previous_n_lower_than_recent = ohlc_with_indicators_daily["lower_than_recent"][-11:-1].tolist()
-    upper_condition = not (False in previous_n_lower_than_recent)
-
-    if output:
-        print(
-            f"- MRI: daily [{format_bool(daily_condition_td)}] / weekly [{format_bool(weekly_condition_td)}] | "
-            f"Consensio: [{format_bool(ma_consensio)}] | MA rising: [{format_bool(ma_rising)}] | "
-            f"Not overextended: [{format_bool(not_overextended)}] \n"
-            f"- Higher close: [{format_bool(daily_condition_close_higher)}] | "
-            f"Volume condition: [{format_bool(volume_condition)}] | Upper condition: [{format_bool(upper_condition)}] | "
-            f"Last candle is green: [{format_bool(last_candle_is_green)}]"
-        )
-
-    confirmation = [
-        daily_condition_td,
-        weekly_condition_td,
-        ma_consensio,
-        ma_rising,
-        not_overextended,
-        daily_condition_close_higher,
-        volume_condition,
-        upper_condition,
-        last_candle_is_green
-    ]
-    numerical_score = round(5 * sum(confirmation) / len(confirmation), 1)  # score X (of 5)
-    result = False not in confirmation
-
-    return result, numerical_score
 
 
 def generate_indicators_daily_weekly(ohlc_daily):
@@ -178,21 +71,30 @@ def get_industry_momentum():
     return industry_momentum, industry_score
 
 
-def get_test_stocks():
+def report_on_shortlist(shortlist, industry_score):
+    # Get the sectors for shortlisted stocks only
+    print(f"Getting industry data for {len(shortlist)} shortlisted stocks, hold on...")
 
-    class Stk:
-        code, name = None, None
+    # Get stock codes to collect industries
+    stock_codes = [stock[0] for stock in shortlist]
+    sectors = get_industry_from_web_batch(stock_codes)
 
-    test_stock = Stk()
-    test_stock.code = 'TTA'
-    test_stock.name = 'TTA'
-    return [test_stock]
+    print(f"All shortlisted stocks (sorted by 5-day moving average volume):")
+    for stock in shortlist:
+
+        # May not find a sector for a stock
+        if sectors[stock[0]] == '-':
+            print(f"- {stock[0]} ({stock[1]}) | {format_number(stock[2])} vol | "
+                  f"Sector score unavailable")
+        else:
+            industry_code = industry_mapping[sectors[stock[0]]]
+            print(f"- {stock[0]} ({stock[1]}) | {format_number(stock[2])} vol | "
+                  f"{sectors[stock[0]]} score {industry_score[industry_code]}/5")
 
 
 def scan_stocks():
     shortlisted_stocks = []
-    #stocks = get_stocks(price_min=price_min, price_max=price_max)
-    stocks = get_test_stocks()
+    stocks = get_stocks(price_min=price_min, price_max=price_max)
 
     # Limit per arguments as required
     if arguments['num'] is not None:
@@ -211,34 +113,33 @@ def scan_stocks():
 
         if ohlc_daily is None:
             print("No data on the asset")
-            continue  # skip if no data
+            continue  # skip this asset if there is no data
 
-        if ohlc_daily is not None:
-            ohlc_with_indicators_daily, ohlc_with_indicators_weekly = generate_indicators_daily_weekly(ohlc_daily)
-            if ohlc_with_indicators_daily is None or ohlc_with_indicators_weekly is None:
-                continue
+        ohlc_with_indicators_daily, ohlc_with_indicators_weekly = generate_indicators_daily_weekly(ohlc_daily)
+        if ohlc_with_indicators_daily is None or ohlc_with_indicators_weekly is None:
+            continue
 
-            confirmation, _ = met_conditions_bullish(
-                ohlc_with_indicators_daily,
-                volume_daily,
-                ohlc_with_indicators_weekly,
-                consider_volume_spike=True,
-                output=True
-            )
-            if confirmation:
-                print("- [v] meeting shortlisting conditions")
-                volume_MA_5D = last_volume_5D_MA(volume_daily)
+        confirmation, _ = met_conditions_bullish(
+            ohlc_with_indicators_daily,
+            volume_daily,
+            ohlc_with_indicators_weekly,
+            consider_volume_spike=True,
+            output=True
+        )
+        if confirmation:
+            print("- [v] meeting shortlisting conditions")
+            volume_MA_5D = last_volume_5D_MA(volume_daily)
 
-                if volume_MA_5D > minimum_volume_level:
-                    print(f"- [v] meeting minimum volume level conditions "
-                          f"({format_number(volume_MA_5D)} > {format_number(minimum_volume_level)})")
-                    shortlisted_stocks.append((stock.code, stock.name, volume_MA_5D))
-                else:
-                    print(f"- [x] not meeting minimum volume level conditions "
-                          f"({format_number(volume_MA_5D)} < {format_number(minimum_volume_level)})")
-
+            if volume_MA_5D > minimum_volume_level:
+                print(f"- [v] meeting minimum volume level conditions "
+                      f"({format_number(volume_MA_5D)} > {format_number(minimum_volume_level)})")
+                shortlisted_stocks.append((stock.code, stock.name, volume_MA_5D))
             else:
-                print("- [x] not meeting shortlisting conditions")
+                print(f"- [x] not meeting minimum volume level conditions "
+                      f"({format_number(volume_MA_5D)} < {format_number(minimum_volume_level)})")
+
+        else:
+            print("- [x] not meeting shortlisting conditions")
 
     # Sort by volume (index 2) descending
     sorted_stocks = sorted(shortlisted_stocks, key=lambda tup: tup[2], reverse=True)
@@ -246,25 +147,7 @@ def scan_stocks():
 
     print()
     if len(shortlist) > 0:
-
-        # Get the sectors for shortlisted stocks only
-        print(f"Getting industry data for {len(shortlist)} shortlisted stocks, hold on...")
-
-        # Get stock codes to collect industries
-        stock_codes = [stock[0] for stock in shortlist]
-        sectors = get_industry_from_web_batch(stock_codes)
-
-        print(f"All shortlisted stocks (sorted by 5-day moving average volume):")
-        for stock in shortlist:
-
-            # May not find a sector for a stock
-            if sectors[stock[0]] == '-':
-                print(f"- {stock[0]} ({stock[1]}) | {format_number(stock[2])} vol | "
-                  f"Sector score unavailable")
-            else:
-                industry_code = industry_mapping[sectors[stock[0]]]
-                print(f"- {stock[0]} ({stock[1]}) | {format_number(stock[2])} vol | "
-                      f"{sectors[stock[0]]} score {industry_score[industry_code]}/5")
+        report_on_shortlist(shortlist, industry_score)
 
     else:
         print(f"No shortlisted stocks")
